+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * AADGroup
   * Added support for Group Licensing by adding the AssignedLicenses property.
+  * Added support for members and owners.
+  FIXES [#1066](https://github.com/microsoft/Microsoft365DSC/issues/1066)
 * TeamsTenantDialPlan
   * Fixed an issue where the Normalization Rules strings were not properly exited.
     FIXES [#2096](https://github.com/microsoft/Microsoft365DSC/issues/2096)